Signal Processing

Signal Processing

It is expected that these functions will be moved to a different package.

FFTW.filt — Function.

filt(b, a, x, [si])

Apply filter described by vectors a and b to vector x, with an optional initial filter state vector si (defaults to zeros).

source

FFTW.filt! — Function.

filt!(out, b, a, x, [si])

Same as filt but writes the result into the out argument, which may alias the input x to modify it in-place.

source

FFTW.deconv — Function.

deconv(b,a) -> c

Construct vector c such that b = conv(a,c) + r. Equivalent to polynomial division.

source

FFTW.conv — Function.

conv(u,v)

Convolution of two vectors. Uses FFT algorithm.

source

FFTW.conv2 — Function.

conv2(u,v,A)

2-D convolution of the matrix A with the 2-D separable kernel generated by the vectors u and v. Uses 2-D FFT algorithm.

source

conv2(B,A)

2-D convolution of the matrix B with the matrix A. Uses 2-D FFT algorithm.

source

FFTW.xcorr — Function.

xcorr(u,v)

Compute the cross-correlation of two vectors.

source